+++ CHANGELOG.md
@@ -1,9 +1,13 @@
 # Changelog
 
+## 3.9.1
+- Fixed sharing badge to LinkedIn issue
+
+
 ## 3.9.0
 
 ### New features
-- BadgeIssuer plugin
+- badgeIssuer plugin
 - Information about each page
 
 ### Updates